backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
chromadb==0.5.23
sentence-transformers==3.3.1
PyPDF2==3.0.1
langchain==0.3.13
langchain-text-splitters==0.3.4
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic>=2.7.4,<3.0.0
requests==2.32.3